requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
aiohttp==3.9.1
jinja2==3.1.3
slowapi==0.1.9
limits==3.7.0
trimesh[easy]==4.0.5
vtracer==0.6.11
pillow==10.2.0
numpy==1.26.3
scipy==1.11.4
shapely==2.0.2
matplotlib==3.8.2
pydantic==2.5.3
python-dotenv==1.0.0
gunicorn==21.2.0
Authlib==1.3.1
dodopayments==1.84.0
modal==0.62.96
itsdangerous==2.2.0
psycopg2-binary==2.9.9
